requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
flask-cors==4.*
pytest==7.*
